+++ tpl/tplimpl/template_ast_transformers_test.go
@@ -10,7 +10,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package tpl
+package tplimpl
 
 import (
 	"bytes"
