+++ tpl/tplimpl/template_embedded.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package tpl
+package tplimpl
 
 type Tmpl struct {
 	Name string
