+++ tpl/tplimpl/template_func_truncate_test.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package tpl
+package tplimpl
 
 import (
 	"html/template"
